rename pst to piece_square_table
